refactor calendarHeader extract horizontalDatePicker in dedicated component component test
